requirements.txt
# RavenCare Medical Triage System - Dependencies

# Web Framework
flask>=3.0.0

# AI Model APIs
google-genai>=0.2.0
openai>=1.12.0
composio>=0.5.0

# Terminal UI
rich>=13.7.0

# PDF Generation
reportlab>=4.0.0

# Security & Encryption
cryptography>=41.0.0
python-dotenv>=1.0.0

# Utilities
python-dateutil>=2.8.2
requests>=2.31.0

# SSL/TLS Support (for production)
pyopenssl>=23.0.0

